Remove copyrighted Apple icon
The icon was imported through react-scripts that imported react-reports that uses node-notifier that in turn uses terminal-notifier, but not as a dependency, as an included binary. That terminal-notifier has not been updated in over 8 years and is actually not necessary for our use. Making a fork of react-scripts is a poor idea as it would cut us from security updates.
So here, we are simply removing the icon at install time.
If the file is not present, the hook silently ignores it.